% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_unique_by_partition}
\alias{check_unique_by_partition}
\title{Check if data frame columns agree with partitioning}
\usage{
check_unique_by_partition(df, partitioning, columns = NULL)
}
\arguments{
\item{df}{Data frame that shall be checked.}

\item{partitioning}{Nested list, specifying which \code{df} rows belong to one partition.}

\item{columns}{Optional: Only check the specified columns of \code{df}. Defaults to all columns.}
}
\value{
Vector of column names of the columns, that agree with the partitioning.
}
\description{
Efficiently check if the columns of the data frame agree with the partitioning.
}
\details{
Agreement means, that only a unique value is provided per partition per column.
}
