% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_dtu_table}
\alias{create_dtu_table}
\title{Title}
\usage{
create_dtu_table(
  dtu,
  gene_name_info = NULL,
  gene_chromosome_info = NULL,
  gene_description = NULL
)
}
\arguments{
\item{gene_description}{}
}
\value{

}
\description{
Title
}
\seealso{

}
\concept{DTUrtle visualization}
