% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dataframe_to_partitioning}
\alias{dataframe_to_partitioning}
\title{Convert a data frame to a partitioning list.}
\usage{
dataframe_to_partitioning(dataframe)
}
\arguments{
\item{dataframe}{The data frame that shall be converted}
}
\value{
A partitioning of data frame column 1 by data frame column 2.
}
\description{
Convert a data frame to a partitioning list.
}
