% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_html_header}
\alias{get_html_header}
\title{Basic HTML header for DTU table}
\usage{
get_html_header()
}
\value{
A HTML header as as string.
}
\description{
This header is only used for HTML images.
}
