% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getmax}
\alias{getmax}
\title{Return the maximal absolute difference for all transcripts of the procvided gene}
\usage{
getmax(gID, dturtle)
}
\arguments{
\item{gID}{gene-identifier}

\item{dturtle}{dturtle object}
}
\value{
The maximal absolute difference value
}
\description{
Return the maximal absolute difference for all transcripts of the procvided gene
}
