% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{merge_sparse}
\alias{merge_sparse}
\title{Merge two or more sparse matrices.}
\usage{
merge_sparse(tx_mat)
}
\arguments{
\item{tx_mat}{List of sparse matrices that shall be merged.}
}
\value{
Sparse matrix, containing the information of all provided matrices.
}
\description{
Merge two or more sparse matrices.
}
