% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{move_columns_to_front}
\alias{move_columns_to_front}
\title{Reorder columns}
\usage{
move_columns_to_front(df, columns)
}
\arguments{
\item{df}{The data frame that shall be reordered.}

\item{columns}{One or multiple column names that should be moved to the front of the dataframe. Order is kept.}
}
\value{
Data frame with reordered columns.
}
\description{
Reorder the columns of a data frame.
}
\details{
Sets the desired \code{columns} of the dataframe \code{df} to the first columns. Does not change the order of the others.
}
\examples{
move_columns_to_fron(df, c("new_first_column", "new_second_column"))
}
