% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{no_na}
\alias{no_na}
\title{Replaces NA values by 1.}
\usage{
no_na(x)
}
\arguments{
\item{x}{Vector of values (e.g. pvalues).}
}
\value{
Vector of values with NAs replaced.
}
\description{
Replaces NA values by 1.
}
