% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{partitioning_to_dataframe}
\alias{partitioning_to_dataframe}
\title{Convert a partitioning list to a data frame.}
\usage{
partitioning_to_dataframe(partitioning)
}
\arguments{
\item{partitioning}{The partitioning that shall be converted}
}
\value{
A data frame with two columns, \code{tx} and \code{gene}.
}
\description{
Convert a partitioning list to a data frame.
}
