% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_dtu_table}
\alias{plot_dtu_table}
\title{Title}
\usage{
plot_dtu_table(
  dtu,
  txdf,
  title,
  folder,
  cores = 1,
  image_folder = "images/",
  dtu_table_image = F,
  include_bar = F,
  include_heat = F,
  include_txplot = F,
  heat_all_counts = all_counts_dtu,
  heat_mut = mut_table,
  heat_draw_genes = anno_genes,
  txplot_reduce_introns = F,
  txplot_reduce_introns_fill = "grey95"
)
}
\arguments{
\item{txplot_reduce_introns_fill}{}
}
\value{

}
\description{
Title
}
