% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_prop_bar}
\alias{plot_prop_bar}
\title{Title}
\usage{
plot_prop_bar(
  dturtle,
  genes = NULL,
  meta_gene_id = NULL,
  group_colors = NULL,
  dash_color = "red",
  savepath = NULL
)
}
\arguments{
\item{dturtle}{Result object of \code{posthoc_and_stager()}. Must be of class \code{dturtle}.}

\item{genes}{Character vector of genes to plot. If \code{NULL}, defaults to all found significant genes (\code{sig_genes}).}

\item{savepath}{}
}
\value{

}
\description{
Title
}
\seealso{
Other DTUrtle visualization: 
\code{\link{create_dtu_table}()}
}
\concept{DTUrtle visualization}
