% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_transcripts_view}
\alias{plot_transcripts_view}
\title{Title}
\usage{
plot_transcripts_view(
  dturtle,
  genes = NULL,
  gtf,
  genome,
  one_to_one = NULL,
  reduce_introns = T,
  reduce_introns_fill = "grey95",
  reduce_introns_min_size = 50,
  savepath = NULL,
  filename_ext = "_transcripts.png",
  BPPARAM = BiocParallel::SerialParam(),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
