% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{prop_matrix}
\alias{prop_matrix}
\title{Actual computation of proportion matrices}
\usage{
prop_matrix(mtx, partitioning = NULL)
}
\arguments{
\item{mtx}{A (sparse) transcript-level matrix}

\item{partitioning}{A DRIMSeq partitioning list, specifying which transcripts belong to which genes.}
}
\value{
A (sparse) matrix of transcript proportions
}
\description{
Actual computation of proportion matrices
}
