% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ratio_expression_in}
\alias{ratio_expression_in}
\title{Ratio of expressing samples}
\usage{
ratio_expression_in(drim, type)
}
\arguments{
\item{drim}{DRIMSeq object}

\item{type}{Type of the summarization that shall be performed. Options are:
\itemize{
\item \code{'tx'}: Transcript-level expressed-in ratios.
\item \code{'gene'}: Gene-level expressed-in ratios.
}}
}
\value{
Data frame with the expressed-in ratios.
}
\description{
Get ratio of expressing samples per gene/transcript.
}
\details{
Expressing samples are defined as feature expression > 0. Also splits expressing samples by condtion.
}
