% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_support.R
\name{readin_bustools}
\alias{readin_bustools}
\title{Read bustools output}
\usage{
readin_bustools(files)
}
\arguments{
\item{files}{Path to result folder of the bustools call}
}
\value{
sparseMatrix with cell barcodes and feature information
}
\description{
Read in result files for single-cell data quantified with kallisto and bustools.
}
\details{
Adds cell barcodes and feature information to the sparseMatrix
}
