% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rm_version}
\alias{rm_version}
\title{Remove ensembl version number}
\usage{
rm_version(x)
}
\arguments{
\item{x}{Vector of identifiers.}
}
\value{
Vector of identifiers without version numbers.
}
\description{
Remove the version number of ensembl gene/transcript identifiers.
}
\details{
Removes everything beyond the first dot ('.') in each provided identifier.
}
\examples{
rm_version(c("ENSG00000000001.5", "ENST00000000001.2"))
}
