% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{run_posthoc}
\alias{run_posthoc}
\title{Posthoc filtering}
\usage{
run_posthoc(drim, filt)
}
\arguments{
\item{filt}{Threshold}
}
\value{
A filtered \code{sparseDRIMSeq::results()} data frame.
}
\description{
Perform posthoc filtering.
}
\details{
Sets pvalue and adjusted pvalue of 'filtered' elements to 1.
}
