% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{seurat_add_tx2gene}
\alias{seurat_add_tx2gene}
\title{Add tx2gene}
\usage{
seurat_add_tx2gene(seurat_obj, tx2gene)
}
\arguments{
\item{tx2gene}{}
}
\value{
Seurat object with added data.
}
\description{
Add a data frame as feature level metadata to the Seurat active assay.
}
