% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{smallProportionSD}
\alias{smallProportionSD}
\title{Filter out results, whose standard deviation of proportional ratios is below the filter value.}
\usage{
smallProportionSD(drim, filter)
}
\arguments{
\item{drim}{Results object of DRIMSeq statistical computations}

\item{filter}{A filter threshold, stating the minimal standard deviation to keep.}
}
\value{
A boolean vector, stating the elements to dismiss (==True).
}
\description{
Filter out results, whose standard deviation of proportional ratios is below the filter value.
}
