% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{table_percentage_bar}
\alias{table_percentage_bar}
\title{Column formatter for percentages}
\usage{
table_percentage_bar(color1, color2, digits, color_break = 0, ...)
}
\arguments{
\item{color1}{The color to use below the color breakpoint.}

\item{color2}{The color to use for greater or equal values to the color breakpoint.}

\item{digits}{The number of digits to format the numbers to.}

\item{color_break}{The value, where a color switch should hapen.}

\item{...}{Further arguments passed to \code{\link[formattable:style]{style}}.}
}
\value{
A formatter function, to be used by \code{\link[=plot_dtu_table]{plot_dtu_table()}} or \code{\link[formattable:formattable]{formattable}}.
}
\description{
Create a color bar for percentage columns in \code{\link[=plot_dtu_table]{plot_dtu_table()}}.
}
\details{
Shall be used in \code{\link[=plot_dtu_table]{plot_dtu_table()}} or \code{\link[formattable:formattable]{formattable}}.
}
\seealso{
\code{\link[=create_dtu_table]{create_dtu_table()}} and \code{\link[=plot_dtu_table]{plot_dtu_table()}} for table visualization.

Other DTUrtle formatter: 
\code{\link{table_pval_tile}()}
}
\concept{DTUrtle formatter}
