% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{table_pval_tile}
\alias{table_pval_tile}
\title{Column formatter for p-values}
\usage{
table_pval_tile(
  gradient_min_color,
  gradient_max_color,
  gradient_alpha = NULL,
  digits,
  gradient_na_rm = TRUE,
  ...
)
}
\arguments{
\item{gradient_min_color}{The color of the minimal value in the color gradient.}

\item{gradient_max_color}{The color of the maximal value in the color gradient.}

\item{gradient_alpha}{Logical of whether to include alpha channel. NULL to let the function decide by input.}

\item{digits}{The number of digits to format the numbers to. Scientific notation is used.}

\item{gradient_na_rm}{Logical indicating whether to ignore missing values.}

\item{...}{Further arguments passed to \code{\link[formattable:style]{style}}.}
}
\value{
A formatter function, to be used by \code{\link[=plot_dtu_table]{plot_dtu_table()}} or \code{\link[formattable:formattable]{formattable}}.
}
\description{
Gradually color p-value columns in \code{\link[=plot_dtu_table]{plot_dtu_table()}}.
}
\details{
Shall be used in \code{\link[=plot_dtu_table]{plot_dtu_table()}} or \code{\link[formattable:formattable]{formattable}}. The color bar width is log10-scaled.
}
\seealso{
\code{\link[=create_dtu_table]{create_dtu_table()}} and \code{\link[=plot_dtu_table]{plot_dtu_table()}} for table visualization.

Other DTUrtle formatter: 
\code{\link{table_percentage_bar}()}
}
\concept{DTUrtle formatter}
