% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_proportion_matrix}
\alias{get_proportion_matrix}
\title{Compute proportion matrix}
\usage{
get_proportion_matrix(obj, genes = NULL)
}
\arguments{
\item{obj}{(sparse) matrix or \code{dturtle} object.}

\item{genes}{Specify the genes to subset the data to, if a \code{dturtle} object is provided.}
}
\value{
A (sparse) matrix of transcript proportions
}
\description{
Compute a matrix of transcript proportions per gene. Either a (sparse) count matrix or a \code{dturtle} object can be provided.
}
\details{
If a \code{dturtle} object is provided, a list of genes the data shall be subsetted to can optionally be given.
}
