% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtu_analysis.R
\name{import_counts}
\alias{import_counts}
\title{Import quantification results}
\usage{
import_counts(files, type, ...)
}
\arguments{
\item{files}{Vector of files to be imported. Optionally can be named to keep the samples names.}

\item{type}{Type of the quantification data. All tools supported by tximport can be selected. If you have single-cell data, the use of \code{alevin} is proposed.
\itemize{
\item \code{'salmon'}
\item \code{'alevin'}
\item \code{'kallisto'}
\item \code{'rsem'}
\item \code{'stringtie'}
\item \code{'sailfish'}
\item \code{'none'}
}}

\item{...}{Further parameters to the specific tximport call. See \code{tximport()} for available parameters.}
}
\value{
\itemize{
\item For bulk data: A combined count matrix for all specified samples.
\item For single-cell data (\code{type='alevin'}): A list of count matrices per sample. Should be combined and optionally added to a Seurat object with \code{combine_to_matrix()}.
}
}
\description{
Import the quantification results of many RNAseq quantifiers, including 'alevin' for single-cell data.
Most likely the first step in your DTUrtle analysis.
}
\details{
Can perform multiple scaling schemes, defaults to scaling schemes appropriate for DTU analysis.
}
\seealso{
Other DTUrtle: 
\code{\link{combine_to_matrix}()},
\code{\link{posthoc_and_stager}()},
\code{\link{run_drimseq}()}
}
\concept{DTUrtle}
