% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{one_to_one_mapping}
\alias{one_to_one_mapping}
\title{Ensure one-to-one mapping}
\usage{
one_to_one_mapping(name, id, ext = "_")
}
\arguments{
\item{name}{A character vector. If no one-to-one mapping exists, the values of this vector will be changed (by extending with \code{ext})!}

\item{id}{A vector, preferably for identifiers. This column will not be touched in case of a disagreement.}

\item{ext}{The extension character.}
}
\value{
The vector \code{name}, where one to one mapping for the two vectors is ensured.
}
\description{
Ensure one-to-one mapping of the two specified vectors.
}
\details{
First checks if every unique value in \code{name} corresponds with a unique value in \code{id}.
If not, changes the disagreeing values in \code{name} by extending the label with the \code{ext} character and a number.
}
