% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtu_analysis.R
\name{posthoc_and_stager}
\alias{posthoc_and_stager}
\title{Title}
\usage{
posthoc_and_stager(dtu, ofdr = 0.05, posthoc = T, posthoc_filt = 0.1)
}
\arguments{
\item{posthoc_filt}{}
}
\value{

}
\description{
Title
}
