% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{run_posthoc}
\alias{run_posthoc}
\title{Pos-thoc filtering}
\usage{
run_posthoc(drim, filt)
}
\arguments{
\item{drim}{Result object of DRIMSeq statistical tests.}

\item{filt}{Threshold to filter by.}
}
\value{
A filtered \code{sparseDRIMSeq::results()} data frame.
}
\description{
Perform post-hoc filtering.
}
\details{
Sets pvalue and adjusted pvalue of 'filtered' elements to 1.
}
